refactor code and add tests

+++ src/main/java/ru/caloriemanager/service/UserServiceCopy.java
@@ -1,9 +1,7 @@
 package ru.caloriemanager.service;
 
-import jakarta.validation.constraints.Min;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 import ru.caloriemanager.model.AbstractNamedEntity;
 import ru.caloriemanager.model.User;
@@ -12,33 +10,39 @@ import ru.caloriemanager.util.exception.NotFoundException;
 
 import java.util.Comparator;
 import java.util.List;
+
 import static ru.caloriemanager.util.ValidationUtil.*;
 
 @Service
-public class UserService {
+public class UserServiceCopy {
     @Qualifier("inMemoryUserRepo")
     @Autowired
     public UserRepository repository;
 
-    public User create(@NonNull User user) {
+    public User create(User user) {
+        if (user == null) throw new IllegalArgumentException("not valid arguments");
         checkNew(user);
         return repository.save(user);
     }
 
-    public void update(@NonNull User user) throws NotFoundException {
+    public void update(User user) throws NotFoundException {
+        if (user == null) throw new IllegalArgumentException("not valid arguments");
         checkNotFoundWithId(repository.save(user), user.getId());
     }
 
-    public void delete(@Min(1) int id) throws NotFoundException {
+    public void delete(int id) throws NotFoundException {
+        if (id < 1) throw new IllegalArgumentException("not valid arguments");
         boolean isDeleted = repository.delete(id);
         checkNotFoundWithId(isDeleted, id);
     }
 
-    public User get(@Min(1) int id) throws NotFoundException {
+    public User get(int id) throws NotFoundException {
+        if (id < 1) throw new IllegalArgumentException("not valid arguments");
         return checkNotFoundWithId(repository.get(id), id);
     }
 
-    public User getByEmail(@NonNull String email) throws NotFoundException {
+    public User getByEmail(String email) throws NotFoundException {
+        if (email == null) throw new IllegalArgumentException("not valid arguments");
         return checkNotFound(repository.getByEmail(email), "email=" + email);
     }
